refactor(server): Split subscription/manager.rs into focused modules
Splits the 2,180-line manager.rs file into 7 focused modules:

- mod.rs (119 lines): Core SubscriptionManager struct and construction
- lifecycle.rs (349 lines): subscribe/unsubscribe operations
- wire.rs (259 lines): Wire protocol integration (UUID bridging, result updates)
- query.rs (358 lines): Query execution and retry logic with backoff
- events.rs (126 lines): Change event handling and notification
- metrics.rs (69 lines): Metrics and configuration accessors
- tests.rs (960 lines): Comprehensive test suite

Each module handles a single concern with no circular dependencies.
All 33 existing tests continue to pass.

Closes #4125
